+++ example/ResetValidatorExample.php
@@ -2,10 +2,13 @@
 
 require __DIR__.'/../vendor/autoload.php';
 
+use LDL\Framework\Base\Contracts\ArrayFactoryInterface;
+use LDL\Validators\Chain\Item\ValidatorChainItem;
 use LDL\Validators\Chain\OrValidatorChain;
-use LDL\Validators\Config\BasicValidatorConfig;
+use LDL\Validators\Config\Traits\ValidatorConfigTrait;
 use LDL\Validators\Config\ValidatorConfigInterface;
 use LDL\Validators\RegexValidator;
+use LDL\Validators\Traits\NegatedValidatorTrait;
 use LDL\Validators\ValidatorInterface;
 use LDL\Validators\ResetValidatorInterface;
 use LDL\Validators\Traits\ValidatorValidateTrait;
@@ -13,21 +16,41 @@ use LDL\Validators\HasValidatorResultInterface;
 use LDL\Validators\Traits\ValidatorHasConfigInterfaceTrait;
 use LDL\Validators\Traits\ValidatorDescriptionTrait;
 
+class ResetValidatorExampleConfig implements ValidatorConfigInterface
+{
+    use ValidatorConfigTrait;
+
+    public static function fromArray(array $data = []): ArrayFactoryInterface
+    {
+        return new self();
+    }
+
+    public function toArray(): array
+    {
+        return [];
+    }
+
+}
+
 class ResetValidatorExample implements ValidatorInterface, HasValidatorResultInterface, ResetValidatorInterface
 {
     use ValidatorValidateTrait;
+    use NegatedValidatorTrait;
     use ValidatorHasConfigInterfaceTrait;
     use ValidatorDescriptionTrait;
+
+    private const DESCRIPTION = 'Reset validator';
 
     /**
      * @var int
      */
     private $internalState;
 
-    public function __construct(bool $negated=false, bool $dumpable=true, string $description=null)
+    public function __construct(bool $negated=false, string $description=null)
     {
-        $this->_tConfig = new BasicValidatorConfig($negated, $dumpable);
-        $this->_tDescription = $description;
+        $this->_tNegated = $negated;
+        $this->_tDescription = $description ?? self::DESCRIPTION;
+        $this->_tConfig = new ResetValidatorExampleConfig();
     }
 
     public function reset()
@@ -53,9 +76,9 @@ class ResetValidatorExample implements ValidatorInterface, HasValidatorResultInt
     {
     }
 
-    public static function fromConfig(ValidatorConfigInterface $config, string $description=null): ValidatorInterface
+    public static function fromConfig(ValidatorConfigInterface $config, bool $negated = false, string $description=null): ValidatorInterface
     {
-        return new self($config->isNegated(), $config->isDumpable(), $description);
+        return new self($negated, $description);
     }
 }
 
@@ -73,14 +96,14 @@ echo "Validate: 'world'\n";
 $chain->validate('world');
 
 echo "Append ResetValidatorExample\n";
-$chain->append(new ResetValidatorExample());
+$chain->append(new ValidatorChainItem(new ResetValidatorExample()));
 
 echo "Validate: 123\n";
 $chain->validate(123);
 echo "OK!\n";
 
 echo "Checking internal state\n";
-$firstValue = $chain->getLast()->getResult();
+$firstValue = $chain->getLast()->getValidator()->getResult();
 echo $firstValue."\n";
 
 echo "Validate: 456\n";
@@ -88,7 +111,7 @@ $chain->validate(456);
 echo "OK!\n";
 
 echo "Checking internal state\n";
-$secondValue = $chain->getLast()->getResult();
+$secondValue = $chain->getLast()->getValidator()->getResult();
 echo $secondValue."\n";
 
 if($firstValue === $secondValue){