+++ src/LDL/Validators/Chain/Dumper/ValidatorChainPhpDumper.php
@@ -2,22 +2,31 @@
 
 namespace LDL\Validators\Chain\Dumper;
 
+use LDL\Validators\Chain\Item\ValidatorChainItemInterface;
 use LDL\Validators\Chain\ValidatorChainInterface;
 use LDL\Validators\ValidatorHasConfigInterface;
 
 class ValidatorChainPhpDumper implements ValidatorChainDumperInterface
 {
-    public static function dump(ValidatorChainInterface $collection) : array
+    public static function dump(ValidatorChainInterface $chain) : array
     {
-        if(count($collection) === 0){
+        if(count($chain) === 0){
             return [];
         }
 
         $data = [
-            'parent' => get_class($collection)
+            'parent' => get_class($chain)
         ];
 
-        foreach($collection->filterDumpableItems() as $validator){
+        if(!$chain instanceof FilterDumpableInterface){
+            return $data;
+        }
+
+        /**
+         * @var ValidatorChainItemInterface $chainItem
+         */
+        foreach($chain->filterDumpableItems() as $chainItem){
+            $validator = $chainItem->getValidator();
             $temp = [
                 'class' => get_class($validator),
             ];