+++ src/LDL/Validators/Chain/Item/ValidatorChainItemInterface.php
@@ -1,0 +1,18 @@
+<?php declare(strict_types=1);
+
+namespace LDL\Validators\Chain\Item;
+
+use LDL\Validators\ValidatorInterface;
+
+interface ValidatorChainItemInterface
+{
+    /**
+     * @return bool
+     */
+    public function isDumpable() : bool;
+
+    /**
+     * @return ValidatorInterface
+     */
+    public function getValidator() : ValidatorInterface;
+}