+++ src/LDL/Validators/Config/Traits/ValidatorConfigTrait.php
@@ -2,19 +2,8 @@
 
 namespace LDL\Validators\Config\Traits;
 
-
 trait ValidatorConfigTrait
 {
-    /**
-     * @var bool
-     */
-    private $_tDumpable;
-
-    public function isDumpable() : bool
-    {
-        return $this->_tDumpable;
-    }
-
     public function jsonSerialize() : array
     {
         return $this->toArray();