+++ src/LDL/Validators/EmailValidator.php
@@ -3,21 +3,23 @@
 namespace LDL\Validators;
 
 use LDL\Validators\Exception\TypeMismatchException;
+use LDL\Validators\Traits\NegatedValidatorTrait;
 use LDL\Validators\Traits\ValidatorDescriptionTrait;
 use LDL\Validators\Traits\ValidatorHasConfigInterfaceTrait;
 use LDL\Validators\Traits\ValidatorValidateTrait;
 
-class EmailValidator implements ValidatorInterface, NegatedValidatorInterface, ValidatorHasConfigInterface
+class EmailValidator implements ValidatorInterface, NegatedValidatorInterface
 {
     use ValidatorValidateTrait;
+    use NegatedValidatorTrait;
     use ValidatorHasConfigInterfaceTrait;
     use ValidatorDescriptionTrait;
 
     private const DESCRIPTION = 'Validate email';
 
-    public function __construct(bool $negated=false, bool $dumpable=true, string $description=null)
+    public function __construct(bool $negated=false, string $description=null)
     {
-        $this->_tConfig = new Config\BasicValidatorConfig($negated, $dumpable);
+        $this->_tNegated = $negated;
         $this->_tDescription = $description ?? self::DESCRIPTION;
     }
 
@@ -47,28 +49,5 @@ class EmailValidator implements ValidatorInterface, NegatedValidatorInterface, V
         );
 
         throw new TypeMismatchException($msg);
-    }
-
-    /**
-     * @param Config\ValidatorConfigInterface $config
-     * @param string|null $description
-     * @return ValidatorInterface
-     * @throws \InvalidArgumentException
-     */
-    public static function fromConfig(Config\ValidatorConfigInterface $config, string $description=null): ValidatorInterface
-    {
-        if(false === $config instanceof Config\BasicValidatorConfig){
-            $msg = sprintf(
-                'Config expected to be %s, config of class %s was given',
-                __CLASS__,
-                get_class($config)
-            );
-            throw new \InvalidArgumentException($msg);
-        }
-
-        /**
-         * @var Config\ValidatorConfigInterface $config
-         */
-        return new self($config->isNegated(), $config->isDumpable(), $description);
     }
 }