+++ src/LDL/Validators/Traits/ValidatorValidateTrait.php
@@ -3,7 +3,7 @@
 namespace LDL\Validators\Traits;
 
 use LDL\Framework\Helper\ClassRequirementHelperTrait;
-use LDL\Validators\Config\NegatedValidatorConfigInterface;
+use LDL\Validators\NegatedValidatorInterface;
 use LDL\Validators\ValidatorInterface;
 
 trait ValidatorValidateTrait
@@ -14,10 +14,8 @@ trait ValidatorValidateTrait
     {
         $this->requireImplements([ValidatorInterface::class]);
 
-        $config = $this->getConfig();
-
-        if($config instanceof NegatedValidatorConfigInterface){
-            $config->isNegated() ? $this->assertFalse($value, ...$params) : $this->assertTrue($value, ...$params);
+        if($this instanceof NegatedValidatorInterface){
+            $this->isNegated() ? $this->assertFalse($value, ...$params) : $this->assertTrue($value, ...$params);
             return;
         }